requirements.txt
accelerate==1.8.1
apex==0.9.10dev
av==15.0.0
bfloat16==1.2.0
bitsandbytes==0.41.0
cog==0.16.8
cupy==13.6.0
data_processing==0.1.4
decord==0.6.0
deepspeed==0.14.4
diffusers==0.35.2
einops==0.8.1
einops_exts==0.0.4
fastapi==0.123.0
fire==0.7.1
flash_attn==2.1.0
flash_mla==1.0.0.dev0
ftfy==6.3.1
fvcore==0.1.5.post20221221
gradio==6.0.1
h5py==3.9.0
horovod==0.28.1
huggingface_hub==0.33.4
hydra-core==1.3.2
kneed==0.8.5
lightning_fabric==2.6.0
lightning_lite==1.8.6
matplotlib==3.7.1
moviepy==1.0.3
networkx==3.4.2
networkx==2.6.3
numpy==2.3.5
nvidia_ml_py==12.575.51
omegaconf==2.3.0
opencv_python==4.10.0.84
packaging==25.0
pandas==1.5.3
peft==0.4.0
Pillow==9.5.0
Pillow==12.0.0
POT==0.9.5
psutil==7.0.0
pydot==4.0.1
pymysql==1.1.2
pytest==9.0.1
python-dotenv==1.2.1
pytorch_lightning==2.6.0
PyYAML==6.0.2
PyYAML==6.0.3
regex==2024.11.6
Requests==2.32.5
rich==14.2.0
ring_flash_attn==0.1.8
rmm==0.0.1
rouge==1.0.1
safetensors==0.7.0
scenedetect==0.6.4
scikit_learn==1.3.2
scipy==1.16.3
seaborn==0.13.2
sentencepiece==0.1.99
setuptools==78.1.1
sglang==0.5.5.post3
shortuuid==1.0.13
SQLAlchemy==2.0.44
sshtunnel==0.4.0
tiktoken==0.9.0
timm==1.0.22
tokenizers==0.22.1
torch_npu==2.8.0
torchmetrics==1.8.2
torchvision==0.15.1+cu118
tqdm==4.67.1
transformers==4.40.0.dev0
typing_extensions==4.15.0
tyro==0.9.35
uvicorn==0.38.0
wandb==0.23.0
wheel==0.45.1
xformers==0.0.33.post1
yt_dlp==2025.11.12
